Specifying s3_region on `S3Storage` CloudStorage
Due to the boto3 issues (https://github.com/boto/boto3/issues/2989) setting the X-Amz-Credential header,
it is recommended to set either the `s3_region` or the `endpoint_url`
field when configuring an S3Storage provider.

This commit adds the `s3_region` field to all documented S3Storage
examples in the documentation.